refactoring image_queue.py move
